consistent naming and fixed create map args
