+++ google-maps/src/web.ts
@@ -6,7 +6,7 @@ import type {
   CameraArgs,
   AddMarkersArgs,
   CapacitorGoogleMapsPlugin,
-  CreateMapOptions,
+  CreateMapArgs,
   CurrentLocArgs,
   DestroyMapArgs,
   IndoorMapArgs,
@@ -16,6 +16,7 @@ import type {
   TrafficLayerArgs,
   RemoveMarkersArgs,
   OnScrollArgs,
+  MapListenerCallback,
 } from './implementation';
 
 export class CapacitorGoogleMapsWeb
@@ -84,7 +85,7 @@ export class CapacitorGoogleMapsWeb
     throw new Error('Method not implemented.');
   }
 
-  async create(options: CreateMapOptions): Promise<void> {
+  async create(options: CreateMapArgs, _callback?: MapListenerCallback): Promise<void> {
     console.log(`Create map: ${options.id}`);
     await this.importGoogleLib(options.apiKey);
     this.maps[options.id] = {
